Scope role update / remove API to admins
Why:

* similarly to f7afc1c75b4ac7e66ca8f621651f3fc54dcacca7 we want to scope
  the updating and removing of roles to be for admins only (via an admin
  user in the UI and via an admin token in an oauth context)

This change addresses the need by:

* moving these endpoints under the admin api controller
